requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
annotated-types==0.6.0
anyio==4.3.0
asgiref==3.7.2
asttokens==2.2.1
async-timeout==4.0.2
attrs==23.2.0
backcall==0.2.0
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.7
dataclasses-json==0.6.4
decorator==5.1.1
distro==1.9.0
Django==4.2.3
django-cors-headers==3.14.0
django-redis==5.3.0
djangorestframework==3.14.0
docopt==0.6.2
exceptiongroup==1.2.1
executing==1.2.0
frozenlist==1.4.1
greenlet==2.0.2
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.4
jedi==0.18.2
jsonpatch==1.33
jsonpointer==2.4
langchain==0.1.16
langchain-community==0.0.34
langchain-core==0.1.45
langchain-openai==0.1.3
langchain-text-splitters==0.0.1
langchain-together==0.1.0
langsmith==0.1.49
marshmallow==3.21.1
multidict==6.0.5
mypy-extensions==1.0.0
numpy==1.25.0
openai==1.23.2
orjson==3.10.1
packaging==23.2
pandas==1.5.3
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
pipreqs==0.4.13
prompt-toolkit==3.0.39
psycopg2-binary==2.9.6
ptyprocess==0.7.0
pure-eval==0.2.2
pydantic==2.7.0
pydantic_core==2.18.1
Pygments==2.15.1
PyJWT==2.8.0
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
PyYAML==6.0.1
redis==5.0.1
regex==2024.4.16
requests==2.31.0
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.29
sqlparse==0.4.4
sseclient-py==1.8.0
stack-data==0.6.2
tabulate==0.9.0
tenacity==8.2.3
tiktoken==0.6.0
together==0.2.11
tqdm==4.66.2
traitlets==5.9.0
typer==0.9.4
typing-inspect==0.9.0
typing_extensions==4.7.1
urllib3==2.0.3
wcwidth==0.2.6
yarg==0.1.9
yarl==1.9.4